Move all JPA converters to a dedicated package
