+++ core/src/main/java/google/registry/persistence/converter/CurrencyToBillingMapUserType.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
