+++ core/src/main/java/google/registry/persistence/converter/StatusValueSetConverter.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import google.registry.model.eppcommon.StatusValue;
 import java.util.Set;
