+++ core/src/main/java/google/registry/persistence/converter/StringSetConverterBase.java
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 
-import google.registry.persistence.StringCollectionDescriptor.StringCollection;
+import google.registry.persistence.converter.StringCollectionDescriptor.StringCollection;
 import java.util.Set;
 import javax.persistence.AttributeConverter;
