+++ core/src/main/java/google/registry/persistence/converter/UpdateAutoTimestampConverter.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import static google.registry.persistence.transaction.TransactionManagerFactory.jpaTm;
