+++ core/src/main/java/google/registry/persistence/converter/ZonedDateTimeConverter.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package google.registry.persistence;
+package google.registry.persistence.converter;
 
 import java.sql.Timestamp;
 import java.time.ZoneId;
